add @classpath file, and rename package to zOSResBuild
