+++ conf/@classpath
@@ -1,0 +1,1 @@
+.:usr/lpp/java/J8.0_64/lib/*.jar:usr/lpp/java/J8.0_64/lib/ext/*.jar:/opt/lpp/IBM/dbb/lib/dbb.core_1.0.6.jar
