+++ src/main/zOS/polycephalyzOSResBuild/CopyFiletoFile.java
@@ -1,4 +1,4 @@
-package com.zos.resbuild;
+package polycephalyzOSResBuild;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
