+++ src/main/zOS/polycephalyzOSResBuild/FileFactoryCopy.java
@@ -1,4 +1,4 @@
-package com.zos.resbuild;
+package polycephalyzOSResBuild;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
