+++ src/main/zOS/polycephalyzOSResBuild/SubmitJob.java
@@ -1,4 +1,4 @@
-package com.zos.resbuild;
+package polycephalyzOSResBuild;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.StringTokenizer;
