+++ src/main/zOS/polycephalyzOSResBuild/package-info.java
@@ -5,4 +5,4 @@
  * @author ge64297
  *
  */
-package com.zos.resbuild;
+package polycephalyzOSResBuild;
